Need Bed, added Archer AI, Added Sprite direction manager, Added redundent scripts because git, Beefed up crosshair, Made game a game
The archer Ai is now in and working, needs work

Added sprites and a controller that faces them left and right and up and down based on rules, don't ask.

Added a bunch of scripts I swear I wrote with Trevor, but never got around to..... Probably still in Git.

As a player you can now fear for your life and murder the innocent.

The warm embrace of death calls to me, or that is just my bed.